update files to main

+++ src/base/ComboBox.java
@@ -184,7 +184,7 @@ public class ComboBox<E> extends JComboBox<E> {
 
         private void createLineStyle(Graphics2D g2) {
             if (isFocusOwner()) {
-                double width = getWidth() - 4;
+                double width = (double) getWidth() - 4;
                 int height = getHeight();
                 g2.setColor(lineColor);
                 double size;
